move the files to the proper folders.
